Mapping is an object, not just a dict.
